Updated auton logic (see longer description)
auton logic is written to have two balls and two hatches loaded securely to acquire 10 points for each set of auton. balls will already be loaded in the necessary spots for cargo autons. it is very likely that the auton script will run for longer than the 15 second sandstorm period.  the driver station will tell the drivers when it is about to run into something, and the drivers, who will be looking at the camera, will be able to use a button press to completely stop the auton loop and drive on their own. all of these options also yield 3 more points for crossing the hab line. all options but the side 1, side 2 option yield 3 points for driving off the level 1 platform. thus, if everything goes accordingly, this auton loop will yield a total of 13 or 16 points. unsure how long each one will take and how much we will have to account for precision in strafing. most likely the only situation in which we might run into another robot during autonomous is if we are both trying to pick up from the same human player. we will have to talk to our alliance about this to determine which auton option is the best.